fixed 64bit compatibility warnings with OSStatus usage in format strings

+++ src/BraintreeRSA.m
@@ -46,7 +46,7 @@
                                   cipherTextBuf, &cipherTextSize);
 
   if(result != noErr)
-    NSLog(@"Secure Key Encryption Failed With Error: %ld", result);
+    NSLog(@"Secure Key Encryption Failed With Error: %ld", (long)result);
 
   return [[NSData alloc] initWithBytes: cipherTextBuf length:cipherTextSize];
 }
